improved label video program, removed useless grasp trackbar

+++ label_video.py
@@ -5,6 +5,7 @@ from grasp import nice_text
 
 
 obj_labels = (
+    'none',
     'bottle',
     'cup',
     'handbag',
@@ -22,7 +23,9 @@ obj_labels = (
 def parse_args() -> argparse.Namespace:
     parser = argparse.ArgumentParser()
     parser.add_argument('video', help='path to video input')
-    parser.add_argument('--output', help='output path', default=None)
+    parser.add_argument('--output', '-o', help='output path', default=None)
+    parser.add_argument('--input-label', '-i', default=None)
+    parser.add_argument('--read-only', '-r', action='store_true', default=False)
     return parser.parse_args()
 
 
@@ -30,6 +33,8 @@ def main() -> None:
     args = parse_args()
     if args.output is None:
         args.output = args.video[:-4] + '_true.csv'
+    if args.input_label is None:
+        args.input_label = args.output
 
     cap = cv2.VideoCapture(args.video)
     orig_frames = []
@@ -43,36 +48,44 @@ def main() -> None:
 
     if len(orig_frames) == 0:
         raise ValueError('Could not read any frame')
+
     frame_idx = 0
-    grasps = [0] * len(orig_frames)
-    labels = [0] * len(orig_frames)
-    hold = 0
-    last_grasp = 0
-    last_label = 0
+    hold = False
+    labels = []
+    input_file = None
+    try:
+        input_file = open(args.input_label, 'r')
+    except FileNotFoundError:
+        labels = [0] * len(orig_frames)
+        last_label = 0
+
+    if input_file is not None:
+        try:
+            reader = csv.DictReader(input_file)
+            for row in reader:
+                labels.append(obj_labels.index(row['label']))
+            last_label = labels[0]
+        finally:
+            input_file.close()
 
     def change_frame(x):
-        nonlocal frame_idx, grasps, labels
+        nonlocal frame_idx, labels
         frame_idx = cv2.getTrackbarPos('frame', f'{args.video}')
-        if hold:
-            grasps[frame_idx] = last_grasp
-            labels[frame_idx] = last_label
-        cv2.setTrackbarPos('grasp', f'{args.video}', grasps[frame_idx])
-        cv2.setTrackbarPos('obj_label', f'{args.video}', labels[frame_idx])
-        draw_image()
-
-    def change_grasp(x):
-        nonlocal grasps
-        grasps[frame_idx] = cv2.getTrackbarPos('grasp', f'{args.video}')
+        if not args.read_only:
+            if hold:
+                labels[frame_idx] = last_label
+            cv2.setTrackbarPos('obj_label', f'{args.video}', labels[frame_idx])
         draw_image()
 
     def change_label(x):
-        nonlocal labels
+        nonlocal labels, last_label
         labels[frame_idx] = cv2.getTrackbarPos('obj_label', f'{args.video}')
+        if hold:
+            last_label = labels[frame_idx]
         draw_image()
 
     def change_hold(x):
-        nonlocal hold, last_grasp, last_label
-        last_grasp = cv2.getTrackbarPos('grasp', f'{args.video}')
+        nonlocal hold, last_label
         last_label = cv2.getTrackbarPos('obj_label', f'{args.video}')
         if cv2.getTrackbarPos('hold', f'{args.video}') == 1:
             hold = True
@@ -81,29 +94,34 @@ def main() -> None:
 
     def draw_image():
         frame = orig_frames[frame_idx].copy()
-        if grasps[frame_idx] == 1:
-            frame = nice_text(frame, 'GRASP', (10, 50))
-        frame = nice_text(frame, f'{obj_labels[labels[frame_idx]]}', (10, 20))
+        if labels[frame_idx] > 0:
+            frame = nice_text(frame, f'{obj_labels[labels[frame_idx]]}', (10, 20))
         cv2.imshow(f'{args.video}', frame)
 
     cv2.namedWindow(f'{args.video}')
-    cv2.createTrackbar('grasp', f'{args.video}', 0, 1, change_grasp)
-    cv2.createTrackbar('obj_label', f'{args.video}', 0, 10, change_label)
     cv2.createTrackbar('frame', f'{args.video}', 0, len(orig_frames) - 1, change_frame)
-    cv2.createTrackbar('hold', f'{args.video}', 0, 1, change_hold)
+    if not args.read_only:
+        cv2.createTrackbar('obj_label', f'{args.video}', 0, 11, change_label)
+        cv2.createTrackbar('hold', f'{args.video}', 0, 1, change_hold)
     draw_image()
+    if args.read_only:
+        print('Press Enter or ctrl + C to terminate')
+    else:
+        print('Press Enter if you want to save, ctrl + C to abort')
     while (cv2.waitKey(1) & 0xFF) != ord('\r'):
         pass
     cv2.destroyAllWindows()
 
-    with open(args.output, 'w') as f:
-        csv_writer = csv.DictWriter(f, ('grasp', 'label'))
-        csv_writer.writeheader()
-        for grasp, label in zip(grasps, labels):
-            csv_writer.writerow(dict(
-                grasp=grasp,
-                label=obj_labels[label],
-            ))
+    if not args.read_only:
+        with open(args.output, 'w') as f:
+            csv_writer = csv.DictWriter(f, ('grasp', 'label'))
+            csv_writer.writeheader()
+            for label in labels:
+                csv_writer.writerow(dict(
+                    grasp=1 if label > 0 else 0,
+                    label=obj_labels[label],
+                ))
+        print('Saved new label file')
 
 
 if __name__ == '__main__':
